Unified `participant-stats`
